// components/enums/page-api-paths-enum.tsx
export enum APIPaths {
    PAGE_INSERT = '/api/page/insert',
    PAGE_UPDATE = '/api/page/update',
    PAGE_DELETE = '/api/page/delete',
    PAGE_GET_ONE = '/api/page/get',
    PAGE_GET_ONE_TABS = '/api/page/getPageTabs',
    PAGE_GET_ALL = '/api/page/getAll',
    MAP_GET_ALL_STUDY = '/api/map/getAllStudy',
    MAP_GET_ALL_REPERTOIRE = '/api/map/getAllRepertoire',
    MAP_GET_ALL_INDEXEA = '/api/map/getAllIndexeA',
    MAP_GET_ALL_INDEXEB = '/api/map/getAllIndexeB',
    USER_UPDATE_TUTORIALS = '/api/user/updateTutorials',
    FOURNISSEURS_GET_ALL = '/api/fournisseurs/getAll',
    FOURNISSEURS_ADD_ONE = '/api/fournisseurs/add',
    FOURNISSEURS_UPDATE_ONE = '/api/fournisseurs/update',
    FOURNISSEURS_DELETE_ONE = '/api/fournisseurs/delete',
    GRAPH_GET_DATA_ALBUM = '/api/graph/getGraphDataAlbum',
    GRAPH_GET_DATA_VOLETA = '/api/graph/getGraphDataVoletA',
    GRAPH_GET_DATA_VOLETB = '/api/graph/getGraphDataVoletB',
    MAP_GET_ENTREPRISE = '/api/map/getEntreprise',
    SHEETS_GET_DATA = '/api/sheets/getData',
    LENGTH_REPERTOIRE = '/api/repertoire/getLength',
}